Update LiveSplit.Yakuza4.asl
To reiterate from the previous commit, I don't think loading actually takes place during taxi fade-out-fade-in until the taxi driving sound plays. After that sound, the IGT stops incrementing and Loads goes to 1. It seems to me that this is the most logical place to remove loads. Doing so also means there doesn't need to be a list of excluded events, the identification of which is hard to do anyway. It's not as visually satisfying, but it's correct.