Title card consistency (Steam)
There was a problem a while back where different people were getting different strings for the title card, and we fixed it by fudging the numbers around until it worked for everyone.

The problem was just that we weren't grabbing the string from its starting point. The string in question is an absolute path to the subtitle file (for example, mine is F:/SteamLibrary/etc.), and different people have different installation directories on their hardware, so that path would obviously be different for every installation. But if we grab the string at its starting point, and check against a substring a certain length from the end of the string, the result will be normalized to be the same for everyone, regardless of the installation path.

(One key assumption I'm making is that the string pointer starts in the same position on every system. It should, and it would be strange if it didn't. But if anyone does experience an issue, it'll be easier to micro-adjust down from the start of the string rather than coming at it from the end.)