Added GOG version support + golf fix
The GOG pointers are all equivalent codepoints to the Steam pointers, to help with maintenance in the future. In general GOG pointers are 0x65900 less than the Steam pointers, but the lower half of the table (FileTimer and down) are 0x30-0x50 bytes less than that.


I realized that the GolfResults pointer on Steam just gives me "par", which obviously doesn't work. Looking at the string in memory, it's actually an incomplete string that was partially wiped from memory at some point prior, and is only fully wiped from memory when golf ends. I couldn't find a reliable way of predicting where the "start" of the incomplete string would be at the time of golf ending, so I couldn't make that work.

During the golf segment, the pointer I've called MusicSlot3 seems to always contain the path to the golf course ambient audio. That string is then invalidated at an appropriate time for the split to occur. If this turns out to not always work, we could probably find an absolute path for the golf course model file, or one of the golf UI elements.