Chapter and fight splits based on Magic Plot Number
There's a particular list of event strings (located at 0x25848E3A8 in Ver.1.12, begins with "boot"). The value we're calling Plot seems to be an index pointing to the current ongoing event in the game from that list. Since most plot-critical events are visited once uniquely per run, this value can be used to determine the progression of the run in a number of valuable ways.

It's worth mentioning that the value is present again at 0x748 rather than 0x730, but that value is one step behind the other for some reason (until you continue from a Game Over, at which time it catches up until you win the fight).

The list of fight splits is only up until the end of Chapter 1 for the time being.