Temporary fix + decomp stuff
Medara had a problem in the Discord on June 28th. After a debug, I did some decomp to assess the problems. The autosplitter seems to be completely fine based on my testing with him, he must have just had a settings issue.

The problem with the autostart was a problem with Paradigm: his value exhibits 213 when I'd expect it to exhibit 212. It needs a larger fix, and I already think I know how I'm going to go about it, but for now this is a patch to get things working.

In addition, I redid some pointers based on the decomp when I figured out what values were coming from where. TitleCard is now much more reliable, as well as EnemyCount. More to come.